Fix issue 21
Fix #21 and split the Text class into several parts.
