+++ src/lang/Text.php
@@ -9,9 +9,12 @@
  */
 namespace phootwork\lang;
 
-use phootwork\lang\inflector\Inflector;
-use phootwork\lang\inflector\InflectorInterface;
+use phootwork\lang\parts\ArrayConversionsPart;
 use phootwork\lang\parts\CheckerPart;
+use phootwork\lang\parts\ComparisonPart;
+use phootwork\lang\parts\InternalPart;
+use phootwork\lang\parts\SearchPart;
+use phootwork\lang\parts\TransformationsPart;
 
 /**
  * Object representation of an immutable String
@@ -19,7 +22,12 @@ use phootwork\lang\parts\CheckerPart;
  * @author gossi
  */
 class Text implements Comparable {
+	use ArrayConversionsPart;
 	use CheckerPart;
+	use ComparisonPart;
+	use SearchPart;
+	use InternalPart;
+	use TransformationsPart;
 
 	/** @var string */
 	private $string;
@@ -28,7 +36,7 @@ class Text implements Comparable {
 	private $encoding;
 
 	/**
-	 * Initializes a String object and assigns both string and encoding properties
+	 * Initializes a String object ad assigns both string and encoding properties
 	 * the supplied values. $string is cast to a string prior to assignment, and if
 	 * $encoding is not specified, it defaults to mb_internal_encoding(). Throws
 	 * an InvalidArgumentException if the first argument is an array or object
@@ -144,77 +152,6 @@ class Text implements Comparable {
 		$end = mb_substr($this->string, $index, $this->length(), $this->encoding);
 
 		return new self($start . $substring . $end);
-	}
-
-	//
-	//
-	// COMPARISON
-	//
-	//
-
-	/**
-	 * Compares this string to another
-	 *
-	 * @param mixed $compare
-	 *
-	 * @return int
-	 *
-	 * @see \phootwork\lang\Comparable::compareTo()
-	 */
-	public function compareTo($compare): int {
-		return $this->compare($compare);
-	}
-
-	/**
-	 * Compares this string to another string, ignoring the case
-	 *
-	 * @param mixed $compare
-	 *
-	 * @return int Return Values:<br>
-	 * 		&lt; 0 if the object is less than comparison<br>
-	 *  	&gt; 0 if the object is greater than comparison<br>
-	 * 		0 if they are equal.
-	 */
-	public function compareCaseInsensitive($compare): int {
-		return $this->compare($compare, 'strcasecmp');
-	}
-
-	/**
-	 * Compares this string to another
-	 *
-	 * @param string|Text $compare string to compare to
-	 * @param callable $callback
-	 *
-	 * @return int
-	 */
-	public function compare($compare, callable $callback = null): int {
-		if ($callback === null) {
-			$callback = 'strcmp';
-		}
-
-		return $callback($this->string, (string) $compare);
-	}
-
-	/**
-	 * Checks whether the string and the given object are equal
-	 *
-	 * @param mixed $string
-	 *
-	 * @return bool
-	 */
-	public function equals($string): bool {
-		return $this->compareTo($string) === 0;
-	}
-
-	/**
-	 * Checks whether the string and the given object are equal ignoring the case
-	 *
-	 * @param mixed $string
-	 *
-	 * @return bool
-	 */
-	public function equalsIgnoreCase($string): bool {
-		return $this->compareCaseInsensitive($string) === 0;
 	}
 
 	//
@@ -360,177 +297,6 @@ class Text implements Comparable {
 		$end = $this->substring($offset + $length);
 
 		return new self($start . $replacement . $end);
-	}
-
-	//
-	//
-	// SEARCH
-	//
-	//
-
-	/**
-	 * Returns the character at the given zero-related index
-	 *
-	 * <code>
-	 * $str = new Text('Hello World!');<br>
-	 * $str->at(6); // W
-	 *
-	 * $str = new Text('いちりんしゃ');<br>
-	 * $str->at(4) // し
-	 * </code>
-	 *
-	 * @param int $index zero-related index
-	 *
-	 * @return string the found character
-	 */
-	public function at(int $index): string {
-		return mb_substr($this->string, $index, 1, $this->encoding);
-	}
-
-	/**
-	 * Returns an ArrayObject consisting of the characters in the string.
-	 *
-	 * @return ArrayObject An ArrayObject of all chars
-	 */
-	public function chars(): ArrayObject {
-		$chars = new ArrayObject();
-		for ($i = 0, $l = $this->length(); $i < $l; $i++) {
-			$chars->append($this->at($i));
-		}
-
-		return $chars;
-	}
-
-	/**
-	 * Returns the index of a given string, starting at the optional zero-related offset
-	 *
-	 * @param string|Text $string
-	 * @param int $offset zero-related offset
-	 *
-	 * @return int|null int for the index or null if the given string doesn't occur
-	 */
-	public function indexOf($string, int $offset = 0): ?int {
-		$offset = $this->prepareOffset($offset);
-		if ($string == '') {
-			return $offset;
-		}
-		$output = mb_strpos($this->string, (string) $string, $offset, $this->encoding);
-
-		return false === $output ? null : $output;
-	}
-
-	/**
-	 * Returns the last index of a given string, starting at the optional offset
-	 *
-	 * @param string $string
-	 * @param int $offset
-	 *
-	 * @return int|null int for the index or null if the given string doesn't occur
-	 */
-	public function lastIndexOf(string $string, ?int $offset = null): ?int {
-		if (null === $offset) {
-			$offset = $this->length();
-		} else {
-			$offset = $this->prepareOffset($offset);
-		}
-
-		if ($string === '') {
-			return $offset;
-		}
-
-		/* Converts $offset to a negative offset as strrpos has a different
-		 * behavior for positive offsets. */
-		$output = mb_strrpos($this->string, (string) $string, $offset - $this->length(), $this->encoding);
-
-		return false === $output ? null : $output;
-	}
-
-	/**
-	 * Checks whether the string starts with the given string. Case sensitive!
-	 *
-	 * @see Text::startsWithIgnoreCase()
-	 *
-	 * @param string|Text $substring The substring to look for
-	 *
-	 * @return bool
-	 */
-	public function startsWith($substring): bool {
-		$substringLength = mb_strlen((string) $substring, $this->encoding);
-		$startOfStr = mb_substr($this->string, 0, $substringLength, $this->encoding);
-
-		return (string) $substring === $startOfStr;
-	}
-
-	/**
-	 * Checks whether the string starts with the given string. Ignores case.
-	 *
-	 * @see Text::startsWith()
-	 *
-	 * @param string|Text $substring The substring to look for
-	 *
-	 * @return bool
-	 */
-	public function startsWithIgnoreCase($substring): bool {
-		$substring = mb_strtolower((string) $substring, $this->encoding);
-		$substringLength = mb_strlen($substring, $this->encoding);
-		$startOfStr = mb_strtolower(mb_substr($this->string, 0, $substringLength, $this->encoding));
-
-		return (string) $substring === $startOfStr;
-	}
-
-	/**
-	 * Checks whether the string ends with the given string. Case sensitive!
-	 *
-	 * @see Text::endsWithIgnoreCase()
-	 *
-	 * @param string|Text $substring The substring to look for
-	 *
-	 * @return bool
-	 */
-	public function endsWith($substring): bool {
-		$substringLength = mb_strlen((string) $substring, $this->encoding);
-		$endOfStr = mb_substr($this->string, $this->length() - $substringLength, $substringLength, $this->encoding);
-
-		return (string) $substring === $endOfStr;
-	}
-
-	/**
-	 * Checks whether the string ends with the given string. Ingores case.
-	 *
-	 * @see Text::endsWith()
-	 *
-	 * @param string|Text $substring The substring to look for
-	 *
-	 * @return bool
-	 */
-	public function endsWithIgnoreCase($substring): bool {
-		$substring = mb_strtolower((string) $substring, $this->encoding);
-		$substringLength = mb_strlen($substring, $this->encoding);
-		$endOfStr = mb_strtolower(mb_substr($this->string, $this->length() - $substringLength, $substringLength, $this->encoding));
-
-		return (string) $substring === $endOfStr;
-	}
-
-	/**
-	 * Checks whether the given string occurs
-	 *
-	 * @param string|Text $string
-	 *
-	 * @return bool
-	 */
-	public function contains($string): bool {
-		return $this->indexOf($string) !== null;
-	}
-
-	/**
-	 * Performs a regular expression matching with the given regexp
-	 *
-	 * @param string $regexp
-	 *
-	 * @return bool
-	 */
-	public function match(string $regexp): bool {
-		return (bool) preg_match($regexp, $this->string);
 	}
 
 	//
@@ -736,284 +502,6 @@ class Text implements Comparable {
 		return $this->substring(0, $length)->append($substring);
 	}
 
-	//
-	//
-	// ARRAY CONVERSIONS
-	//
-	//
-
-	/**
-	 * Splits the string by string
-	 *
-	 * @param string $delimiter The boundary string.
-	 * @param int $limit
-	 * 		If limit is set and positive, the returned array will contain a maximum of
-	 * 		limit elements with the last element containing the rest of string.
-	 *
-	 * 		If the limit parameter is negative, all components except the last
-	 * 		-limit are returned.
-	 *
-	 * 		If the limit parameter is zero, then this is treated as 1.
-	 *
-	 * @throws \InvalidArgumentException If the delimiter is an empty string.
-	 *
-	 * @return ArrayObject
-	 * 		Returns an array of strings created by splitting the string parameter on boundaries
-	 * 		formed by the delimiter.
-	 *
-	 *        If delimiter contains a value that is not contained in string and a negative limit is used,
-	 *        then an empty array will be returned, otherwise an array containing string will be returned.
-	 *
-	 */
-	public function split(string $delimiter, int $limit = PHP_INT_MAX): ArrayObject {
-		if ('' === $delimiter) {
-			throw new \InvalidArgumentException("The delimiter can't be an empty string");
-		}
-
-		return new ArrayObject(explode($delimiter, $this->string, $limit));
-	}
-
-	/**
-	 * Join array elements with a string
-	 *
-	 * @param array $pieces The array of strings to join.
-	 * @param string $glue Defaults to an empty string.
-	 * @param string $encoding the desired encoding
-	 *
-	 * @return Text
-	 * 		Returns a string containing a string representation of all the array elements in the
-	 * 		same order, with the glue string between each element.
-	 */
-	public static function join(array $pieces, string $glue = '', ?string $encoding = null): self {
-		return new self(implode($glue, $pieces), $encoding);
-	}
-
-	/**
-	 * Convert the string to an array
-	 *
-	 * @param int $splitLength Maximum length of the chunk.
-	 *
-	 * @throws \InvalidArgumentException If splitLength is less than 1.
-	 *
-	 * @return ArrayObject
-	 * 		If the optional splitLength parameter is specified, the returned array will be
-	 * 		broken down into chunks with each being splitLength in length, otherwise each chunk
-	 * 		will be one character in length.
-	 *      If the split_length length exceeds the length of string, the entire string is returned
-	 *      as the first (and only) array element.
-	 */
-	public function chunk(int $splitLength = 1): ArrayObject {
-		$this->verifyPositive($splitLength, 'The chunk length');
-
-		return new ArrayObject(str_split($this->string, $splitLength));
-	}
-
-	//
-	//
-	// TRANSFORMS
-	//
-	//
-
-	/**
-	 * Transforms the string to lowercase
-	 *
-	 * @return Text
-	 */
-	public function toLowerCase(): self {
-		return new self(mb_strtolower($this->string, $this->encoding), $this->encoding);
-	}
-
-	/**
-	 * Transforms the string to first character lowercased
-	 *
-	 * @return Text
-	 */
-	public function toLowerCaseFirst(): self {
-		$first = $this->substring(0, 1);
-		$rest = $this->substring(1);
-
-		return new self(mb_strtolower((string) $first, $this->encoding) . $rest, $this->encoding);
-	}
-
-	/**
-	 * Transforms the string to uppercase
-	 *
-	 * @return Text
-	 */
-	public function toUpperCase(): self {
-		return new self(mb_strtoupper($this->string, $this->encoding), $this->encoding);
-	}
-
-	/**
-	 * Transforms the string to first character uppercased
-	 *
-	 * @return Text
-	 */
-	public function toUpperCaseFirst(): self {
-		$first = $this->substring(0, 1);
-		$rest = $this->substring(1);
-
-		return new self(mb_strtoupper((string) $first, $this->encoding) . $rest, $this->encoding);
-	}
-
-	/**
-	 * Transforms the string to only its first character capitalized.
-	 *
-	 * @return Text
-	 */
-	public function toCapitalCase(): self {
-		return $this->toLowerCase()->toUpperCaseFirst();
-	}
-
-	/**
-	 * Transforms the string with the words capitalized.
-	 *
-	 * @return Text
-	 */
-	public function toCapitalCaseWords(): self {
-		$encoding = $this->encoding;
-
-		return $this->split(' ')->map(function (string $str) use ($encoding) {
-			return self::create($str, $encoding)->toCapitalCase();
-		})->join(' ');
-	}
-
-	/**
-	 * Converts this string into camelCase. Numbers are considered as part of its previous piece.
-	 *
-	 * <code>
-	 * $var = new Text('my_own_variable');<br>
-	 * $var->toCamelCase(); // myOwnVariable
-	 *
-	 * $var = new Text('my_test3_variable');<br>
-	 * $var->toCamelCase(); // myTest3Variable
-	 * </code>
-	 *
-	 * @return Text
-	 */
-	public function toCamelCase(): self {
-		return $this->toStudlyCase()->toLowerCaseFirst();
-	}
-
-	/**
-	 * Converts this string into snake_case. Numbers are considered as part of its previous piece.
-	 *
-	 * <code>
-	 * $var = new Text('myOwnVariable');<br>
-	 * $var->toSnakeCase(); // my_own_variable
-	 *
-	 * $var = new Text('myTest3Variable');<br>
-	 * $var->toSnakeCase(); // my_test3_variable
-	 * </code>
-	 *
-	 * @return Text
-	 */
-	public function toSnakeCase(): self {
-		return $this->toKebabCase()->replace('-', '_');
-	}
-
-	/**
-	 * Converts this string into StudlyCase. Numbers are considered as part of its previous piece.
-	 *
-	 * <code>
-	 * $var = new Text('my_own_variable');<br>
-	 * $var->toStudlyCase(); // MyOwnVariable
-	 *
-	 * $var = new Text('my_test3_variable');<br>
-	 * $var->toStudlyCase(); // MyTest3Variable
-	 * </code>
-	 *
-	 * @return Text
-	 *
-	 * @psalm-suppress InvalidArgument argument 2 of preg_replace_callback CAN BE closure, too (see https://www.php.net/manual/en/function.preg-replace-callback.php)
-	 */
-	public function toStudlyCase(): self {
-		$input = $this->trim('-_');
-		if ($input->isEmpty()) {
-			return $input;
-		}
-		$encoding = $this->encoding;
-
-		return self::create(preg_replace_callback('/([A-Z-_][a-z0-9]+)/', function (array $matches) use ($encoding) {
-			return self::create($matches[0], $encoding)->replace(['-', '_'], '')->toUpperCaseFirst();
-		}, $input->toString()), $this->encoding)->toUpperCaseFirst();
-	}
-
-	/**
-	 * Convert this string into kebab-case. Numbers are considered as part of its previous piece.
-	 *
-	 * <code>
-	 * $var = new Text('myOwnVariable');<br>
-	 * $var->toKebapCase(); // my-own-variable
-	 *
-	 * $var = new Text('myTest3Variable');<br>
-	 * $var->toKebabCase(); // my-test3-variable
-	 * </code>
-	 *
-	 * @return Text
-	 */
-	public function toKebabCase(): self {
-		if ($this->contains('_')) {
-			return $this->replace('_', '-');
-		}
-
-		return new self(mb_strtolower(preg_replace('/([a-z0-9])([A-Z])/', '$1-$2', $this->string)), $this->encoding);
-	}
-
-	/**
-	 * Get the plural form of the Text object.
-	 *
-	 * @param InflectorInterface|null $pluralizer
-	 *
-	 * @return Text
-	 */
-	public function toPlural(?InflectorInterface $pluralizer = null): self {
-		$pluralizer = $pluralizer ?: new Inflector();
-
-		return new self($pluralizer->getPluralForm($this->string), $this->encoding);
-	}
-
-	/**
-	 * Get the singular form of the Text object.
-	 *
-	 * @param InflectorInterface|null $pluralizer
-	 *
-	 * @return Text
-	 */
-	public function toSingular(?InflectorInterface $pluralizer = null): self {
-		$pluralizer = $pluralizer ?: new Inflector();
-
-		return new self($pluralizer->getSingularForm($this->string), $this->encoding);
-	}
-
-	/**
-	 * Converts each tab in the string to some number of spaces, as defined by
-	 * $tabLength. By default, each tab is converted to 4 consecutive spaces.
-	 *
-	 * @param int $tabLength Number of spaces to replace each tab with
-	 *
-	 * @return Text text with tabs converted to spaces
-	 */
-	public function toSpaces(int $tabLength = 4): self {
-		$spaces = str_repeat(' ', $tabLength);
-
-		return $this->replace("\t", $spaces);
-	}
-	/**
-	 * Converts each occurrence of some consecutive number of spaces, as
-	 * defined by $tabLength, to a tab. By default, each 4 consecutive spaces
-	 * are converted to a tab.
-	 *
-	 * @param int $tabLength Number of spaces to replace with a tab
-	 *
-	 * @return Text text with spaces converted to tabs
-	 */
-	public function toTabs(int $tabLength = 4): self {
-		$spaces = str_repeat(' ', $tabLength);
-
-		return $this->replace($spaces, "\t");
-	}
-
 	/**
 	 * Returns the native string
 	 *
@@ -1021,6 +509,10 @@ class Text implements Comparable {
 	 */
 	public function toString(): string {
 		return $this->string;
+	}
+
+	protected function getString(): string {
+		return $this->toString();
 	}
 
 	//
@@ -1031,156 +523,5 @@ class Text implements Comparable {
 
 	public function __toString(): string {
 		return $this->string;
-	}
-
-	//
-	//
-	// INTERNALS
-	//
-	//
-
-	/**
-	 * @internal
-	 *
-	 * @param int $offset
-	 *
-	 * @return int
-	 */
-	protected function prepareOffset(int $offset): int {
-		$len = $this->length();
-		if ($offset < -$len || $offset > $len) {
-			throw new \InvalidArgumentException('Offset must be in range [-len, len]');
-		}
-
-		if ($offset < 0) {
-			$offset += $len;
-		}
-
-		return $offset;
-	}
-
-	/**
-	 * @internal
-	 *
-	 * @param int $offset
-	 * @param int $length
-	 *
-	 * @throws \InvalidArgumentException
-	 *
-	 * @return int
-	 */
-	protected function prepareLength(int $offset, ?int $length): int {
-		if ($length === null) {
-			return $this->length() - $offset;
-		}
-
-		if ($length < 0) {
-			$length += $this->length() - $offset;
-
-			if ($length < 0) {
-				throw new \InvalidArgumentException('Length too small');
-			}
-		} else {
-			if ($offset + $length > $this->length()) {
-				throw new \InvalidArgumentException('Length too large');
-			}
-		}
-
-		return $length;
-	}
-
-	/**
-	 * @internal
-	 *
-	 * @param string|Text $string
-	 * @param string $name
-	 *
-	 * @throws \InvalidArgumentException
-	 */
-	protected function verifyNotEmpty($string, string $name): void {
-		if (empty($string)) {
-			throw new \InvalidArgumentException("$name cannot be empty");
-		}
-	}
-
-	/**
-	 * @internal
-	 *
-	 * @param int $value
-	 * @param string $name
-	 *
-	 * @throws \InvalidArgumentException
-	 */
-	protected function verifyPositive(int $value, string $name): void {
-		if ($value <= 0) {
-			throw new \InvalidArgumentException("$name has to be positive");
-		}
-	}
-
-	/**
-	 * @internal
-	 *
-	 * @param int $value
-	 * @param string $name
-	 *
-	 * @throws \InvalidArgumentException
-	 */
-	protected function verifyNotNegative(int $value, string $name): void {
-		if ($value < 0) {
-			throw new \InvalidArgumentException("$name can not be negative");
-		}
-	}
-
-	/**
-	 * @todo never used: remove?
-	 *
-	 * @internal
-	 *
-	 * @param mixed $replacements
-	 * @param int $limit
-	 *
-	 * @return string
-	 */
-	protected function replacePairs($replacements, ?int $limit): string {
-		if ($limit === null) {
-			return strtr($this->string, $replacements);
-		}
-
-		$this->verifyPositive($limit, 'Limit');
-		$str = $this->string;
-		foreach ($replacements as $from => $to) {
-			$str = $this->replaceWithLimit($str, $from, $to, $limit);
-			if (0 === $limit) {
-				break;
-			}
-		}
-
-		return $str;
-	}
-
-	/**
-	 * @todo never used: remove?
-	 *
-	 * @param Text $str
-	 * @param Text $from
-	 * @param Text $to
-	 * @param int $limit
-	 *
-	 * @return mixed
-	 */
-	protected function replaceWithLimit(self $str, self $from, self $to, int &$limit) {
-		$lenDiff = $to->length() - $from->length();
-		$index = 0;
-
-		while (null !== $index = $str->indexOf($from->toString(), $index)) {
-			$str = $str->splice($to, $index, $to->length());
-			$index += $lenDiff;
-
-			if (0 === --$limit) {
-				break;
-			}
-		}
-
-		return $str;
 	}
 }
